+++ packages/kit-bg/src/services/ServiceScanQRCode/utils/parseQRCode/handlers/url.ts
@@ -1,5 +1,4 @@
 import { parseUrl } from '@onekeyhq/shared/src/utils/uriUtils';
-
 import { EQRCodeHandlerType } from '@onekeyhq/shared/types/qrCode';
 
 import type { IQRCodeHandler, IUrlValue } from '../type';
