+++ packages/kit/src/components/Password/container/PasswordUpdateContainer.tsx
@@ -4,9 +4,9 @@ import { useIntl } from 'react-intl';
 
 import { Toast } from '@onekeyhq/components';
 import backgroundApiProxy from '@onekeyhq/kit/src/background/instance/backgroundApiProxy';
-import { EPasswordMode } from '@onekeyhq/shared/types/password';
 import { usePasswordModeAtom } from '@onekeyhq/kit-bg/src/states/jotai/atoms';
 import { ETranslations } from '@onekeyhq/shared/src/locale';
+import { EPasswordMode } from '@onekeyhq/shared/types/password';
 
 import PasswordSetup from '../components/PasswordSetup';
