+++ packages/kit/src/provider/Container/AirGapQrcodeDialogContainer/AirGapQrcodeDialogContainer.tsx
@@ -1,6 +1,5 @@
 import { useEffect } from 'react';
 
-import { EQRCodeHandlerNames } from '@onekeyhq/shared/types/qrCode';
 import { SecureQRCodeDialogCancel } from '@onekeyhq/shared/src/errors';
 import { toPlainErrorObject } from '@onekeyhq/shared/src/errors/utils/errorUtils';
 import type { IAppEventBusPayload } from '@onekeyhq/shared/src/eventBus/appEventBus';
@@ -8,6 +7,7 @@ import {
   EAppEventBusNames,
   appEventBus,
 } from '@onekeyhq/shared/src/eventBus/appEventBus';
+import { EQRCodeHandlerNames } from '@onekeyhq/shared/types/qrCode';
 
 import backgroundApiProxy from '../../../background/instance/backgroundApiProxy';
 import { SecureQRToast } from '../../../components/SecureQRToast';
