+++ packages/shared/types/qrCode.ts
@@ -25,3 +25,19 @@ export enum EQRCodeHandlerNames {
   marketDetail = 'marketDetail',
   sendProtection = 'sendProtection',
 }
+
+export const PARSE_HANDLER_NAMES = {
+  all: [
+    EQRCodeHandlerNames.bitcoin,
+    EQRCodeHandlerNames.ethereum,
+    EQRCodeHandlerNames.solana,
+    EQRCodeHandlerNames.walletconnect,
+    EQRCodeHandlerNames.migrate,
+    EQRCodeHandlerNames.animation,
+    EQRCodeHandlerNames.urlAccount,
+    EQRCodeHandlerNames.marketDetail,
+    EQRCodeHandlerNames.sendProtection,
+  ],
+  animation: [EQRCodeHandlerNames.animation],
+  none: [],
+} as Record<string, EQRCodeHandlerNames[]>;
